requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beanie==2.0.0
click==8.2.1
dnspython==2.7.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
lazy-model==0.3.0
loguru==0.7.3
motor==3.7.1
pydantic==2.11.7
pydantic_core==2.33.2
pyinstrument==5.0.3
pymongo==4.14.0
python-dotenv==1.1.1
redis==6.4.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
